rename use-detailed-activity-modal
