+++ client/src/components/Today/hooks/useDetailedActivityModal.ts
